:sparkles: add function for changing user data
